Update _config.yml and post.html
This should provide the fix for the broken links. baseurl: "" was missing and in a previous attempt to fix I messed up some of the html in post.html. I reverted and provided the appropriate fix. (the fix sort of breaks local testing but appears to have fixed the links I can get to on a local page.)