Dev/page link fixes (#17)
* Update _config.yml and post.html

This should provide the fix for the broken links. baseurl: "" was missing and in a previous attempt to fix I messed up some of the html in post.html. I reverted and provided the appropriate fix. (the fix sort of breaks local testing but appears to have fixed the links I can get to on a local page.)

* page-link-fixes-v2

I went and looked at the baseurl in the medium theme template and it was done like this. I also fixed the post html back to how it was. It seems to now have fixed the local testing. Only one way to see if it fixes the live version?