changing how html handles post images (#38)
* changing how html handles post images

adding variation for post images so thumbnails don't look so wonky. Also changed file types to webp

* Update post.html

removed adsense from the post.html on this pr so that it would match main when 'pr37-removing google adsense' gets merged in.

* moved image location to _config.ymnl

Moved image location to _config.yml as it doesn't need to be in each individual post.

* remove image location from posts

Removing image location from posts. it seems i forgot to remove it once it was moved elsehwere.

* fixing a couple of conditional statements and fixing pointers to post settings.

Co-authored-by: Flavius-The-Person <Flavius.The.Person@gmail.com>
Co-authored-by: OGJunkyard <ogjunkyard@gmail.com>